refactor(next/parser): extract variable scope analysis from SparqlAstBuilder
