+++ src/main/java/fr/inria/corese/core/next/query/impl/parser/SparqlAstBuilder.java
@@ -3,7 +3,6 @@ package fr.inria.corese.core.next.query.impl.parser;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -39,7 +38,6 @@ import fr.inria.corese.core.next.query.impl.sparql.ast.UnionAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.VarAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.AddAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.AndAst;
-import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.BinaryConstraintAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.BinaryRegexAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.BooleanNotAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.BoundAst;
@@ -63,7 +61,6 @@ import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.SameTermAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.StrAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.SubtractAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.TrinaryRegexAst;
-import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.UnaryConstraintAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.UnaryMinusAst;
 import fr.inria.corese.core.next.query.impl.sparql.ast.constraint.UnaryPlusAst;
 
@@ -149,6 +146,11 @@ public final class SparqlAstBuilder {
      * DESCRIBE resources (IRIs or variables). Set by DescribeQueryFeature.
      */
     private final List<TermAst> describeResources = new ArrayList<>();
+
+    /**
+     * Helper used to compute visible and referenced variables.
+     */
+    private final VariableScopeAnalyzer variableScopeAnalyzer = new VariableScopeAnalyzer();
 
     public SparqlAstBuilder(SparqlParserOptions options) {
         this.options = options;
@@ -335,19 +337,10 @@ public final class SparqlAstBuilder {
             throw new IllegalStateException("No WHERE clause: did you call exitGroup() for the top-level GroupGraphPattern?");
         }
         return switch (this.queryType) {
-            case ASK -> new AskQueryAst(whereClause);
-            case CONSTRUCT -> {
-                // TODO #306: validate variable visibility for solution modifiers when ConstructQueryAst carries them.
-                yield null;
-            }
-            case DESCRIBE -> {
-                // TODO #306: validate variable visibility for solution modifiers when DescribeQueryAst carries them.
-                yield new DescribeQueryAst(describeResources, whereClause);
-            }
-            case SELECT -> {
-                validateSelectQueryVariableVisibility();
-                yield new SelectQueryAst(projection, whereClause, buildSolutionModifier());
-            }
+            case ASK -> buildAskQueryAst();
+            case CONSTRUCT -> buildConstructQueryAst();
+            case DESCRIBE -> buildDescribeQueryAst();
+            case SELECT -> buildSelectQueryAst();
             case UNDEFINED ->
                     throw new QueryEvaluationException("Could not determine the type of query during parsing");
         };
@@ -380,22 +373,48 @@ public final class SparqlAstBuilder {
     }
 
     /**
-     * Validates variable visibility for the parts of SELECT already supported by the next parser.
+     * Builds the AST for ASK queries.
      */
-    private void validateSelectQueryVariableVisibility() {
-        // TODO #306: extend this validation to GROUP BY when it is supported by the next parser.
-        Set<String> visibleVariables = collectVisibleVariables(whereClause);
+    private AskQueryAst buildAskQueryAst() {
+        return new AskQueryAst(whereClause);
+    }
 
-        if (projection.selectAll()) {
-            // SELECT * still needs ORDER BY validation.
-            validateOrderVariables(visibleVariables);
-            return;
+    /**
+     * Builds the AST for SELECT queries.
+     */
+    private SelectQueryAst buildSelectQueryAst() {
+        validateSelectQueryScope();
+        return new SelectQueryAst(projection, whereClause, buildSolutionModifier());
+    }
+
+    /**
+     * Builds the AST for DESCRIBE queries.
+     */
+    private DescribeQueryAst buildDescribeQueryAst() {
+        // TODO #306: validate variable scope for DESCRIBE modifiers when DescribeQueryAst carries them.
+        return new DescribeQueryAst(describeResources, whereClause);
+    }
+
+    /**
+     * Builds the AST for CONSTRUCT queries.
+     */
+    private QueryAst buildConstructQueryAst() {
+        // TODO #306: validate variable scope for CONSTRUCT modifiers when ConstructQueryAst carries them.
+        return null;
+    }
+
+    /**
+     * Validates SELECT projection and ORDER BY variables against the WHERE clause scope.
+     */
+    private void validateSelectQueryScope() {
+        // TODO #306: extend this validation to GROUP BY when it is supported by the next parser.
+        Set<String> visibleVariables = variableScopeAnalyzer.collectVisibleVariables(whereClause);
+
+        // SELECT * still needs ORDER BY validation.
+        if (!projection.selectAll()) {
+            validateProjectionVariables(visibleVariables);
         }
 
-        // Explicit projection variables must come from the WHERE clause scope.
-        validateProjectionVariables(visibleVariables);
-
-        // ORDER BY may use visible variables even if they are not projected.
         validateOrderVariables(visibleVariables);
     }
 
@@ -407,8 +426,7 @@ public final class SparqlAstBuilder {
     private void validateProjectionVariables(Set<String> visibleVariables) {
         for (VarAst projectedVar : projection.variables()) {
             if (!visibleVariables.contains(projectedVar.name())) {
-                // Projection variables must be visible before the SELECT AST is created.
-                throw new QuerySyntaxException(buildVariableVisibilityErrorMessage(
+                throw new QuerySyntaxException(buildOutOfScopeVariableMessage(
                         projectedVar.name(),
                         "SELECT projection"));
             }
@@ -422,13 +440,12 @@ public final class SparqlAstBuilder {
      */
     private void validateOrderVariables(Set<String> visibleVariables) {
         for (OrderConditionAst orderCondition : orderConditions) {
-            // ORDER BY expressions may reference several variables.
-            Set<String> referencedVariables = collectReferencedVariables(orderCondition.expression());
+            Set<String> referencedVariables = variableScopeAnalyzer
+                    .collectReferencedVariables(orderCondition.expression());
 
             for (String variableName : referencedVariables) {
                 if (!visibleVariables.contains(variableName)) {
-                    // Each referenced variable must be visible in the WHERE clause.
-                    throw new QuerySyntaxException(buildVariableVisibilityErrorMessage(
+                    throw new QuerySyntaxException(buildOutOfScopeVariableMessage(
                             variableName,
                             "ORDER BY"));
                 }
@@ -436,158 +453,8 @@ public final class SparqlAstBuilder {
         }
     }
 
-    /**
-     * Builds the error message used when a modifier references a variable outside the WHERE scope.
-     *
-     * @param variableName the variable name without {@code ?}
-     * @param modifier the modifier using the variable
-     * @return a user-facing syntax error message
-     */
-    private String buildVariableVisibilityErrorMessage(String variableName, String modifier) {
-        return "Variable ?" + variableName + " used in " + modifier + " is not visible in WHERE clause";
-    }
-
-    /**
-     * Collects variables visible from the current WHERE clause.
-     *
-     * @param whereClause the WHERE clause of the current query
-     * @return the set of visible variable names, without {@code ?} or {@code $}
-     */
-    private Set<String> collectVisibleVariables(GroupGraphPatternAst whereClause) {
-        Set<String> visibleVariables = new LinkedHashSet<>();
-        if (whereClause == null) {
-            return visibleVariables;
-        }
-
-        for (PatternAst pattern : whereClause.patterns()) {
-            collectVisibleVariables(pattern, visibleVariables);
-        }
-
-        return visibleVariables;
-    }
-
-    /**
-     * Recursively collects visible variables from supported graph patterns.
-     *
-     * @param pattern the pattern to inspect
-     * @param visibleVariables the accumulator for visible variable names
-     */
-    private void collectVisibleVariables(PatternAst pattern, Set<String> visibleVariables) {
-        if (pattern == null) {
-            return;
-        }
-
-        switch (pattern) {
-            case BgpAst(List<TriplePatternAst> triples) -> {
-                // Variables come from triple terms.
-                for (TriplePatternAst triple : triples) {
-                    addIfVariable(triple.subject(), visibleVariables);
-                    addIfVariable(triple.predicate(), visibleVariables);
-                    addIfVariable(triple.object(), visibleVariables);
-                }
-            }
-
-            case GroupGraphPatternAst(List<PatternAst> patterns) -> {
-                // Recurse into nested groups.
-                for (PatternAst nestedPattern : patterns) {
-                    collectVisibleVariables(nestedPattern, visibleVariables);
-                }
-            }
-
-            case OptionalAst(PatternAst optionalPattern) ->
-                // OPTIONAL keeps variables in scope.
-                collectVisibleVariables(optionalPattern, visibleVariables);
-
-            case UnionAst(GroupGraphPatternAst left, GroupGraphPatternAst right) -> {
-                // UNION exposes variables from both branches.
-                collectVisibleVariables(left, visibleVariables);
-                collectVisibleVariables(right, visibleVariables);
-            }
-
-            case FilterAst ignored -> {
-                // FILTER does not make a variable visible by itself.
-            }
-        }
-    }
-
-    /**
-     * Adds the term name when the term is a variable.
-     *
-     * @param term the term to inspect
-     * @param visibleVariables the accumulator for visible variable names
-     */
-    private void addIfVariable(TermAst term, Set<String> visibleVariables) {
-        if (term instanceof VarAst(String name)) {
-            visibleVariables.add(name);
-        }
-    }
-
-    /**
-     * Collects variables referenced by a term or expression.
-     *
-     * @param term the term or expression to inspect
-     * @return the set of referenced variable names
-     */
-    private Set<String> collectReferencedVariables(TermAst term) {
-        Set<String> referencedVariables = new LinkedHashSet<>();
-        collectReferencedVariables(term, referencedVariables);
-        return referencedVariables;
-    }
-
-    /**
-     * Recursively collects variables referenced by a term or expression.
-     *
-     * @param term the term or expression to inspect
-     * @param referencedVariables the accumulator for referenced variable names
-     */
-    private void collectReferencedVariables(TermAst term, Set<String> referencedVariables) {
-        if (term == null) {
-            return;
-        }
-
-        switch (term) {
-            case VarAst(String name) -> {
-                // A direct variable reference contributes to ORDER BY validation.
-                referencedVariables.add(name);
-            }
-
-            case UnaryConstraintAst unaryConstraint -> {
-                // Recurse into unary expressions such as STR(?x) or BOUND(?x).
-                collectReferencedVariables(unaryConstraint.getArgument(), referencedVariables);
-            }
-
-            case BinaryConstraintAst binaryConstraint -> {
-                // Recurse into both operands of binary expressions.
-                collectReferencedVariables(binaryConstraint.getLeftArgument(), referencedVariables);
-                collectReferencedVariables(binaryConstraint.getRightArgument(), referencedVariables);
-            }
-
-            case FunctionCallAst(TermAst ignored, List<TermAst> arguments) -> {
-                // Recurse into each function argument.
-                for (TermAst argument : arguments) {
-                    collectReferencedVariables(argument, referencedVariables);
-                }
-            }
-
-            case TrinaryRegexAst regexAst -> {
-                // REGEX may reference variables in the text, pattern or flags.
-                collectReferencedVariables(regexAst.getString(), referencedVariables);
-                collectReferencedVariables(regexAst.getPattern(), referencedVariables);
-                collectReferencedVariables(regexAst.getFlags(), referencedVariables);
-            }
-
-            case IriAst ignoredIri -> {
-                // Constants do not contribute referenced variables.
-            }
-
-            case LiteralAst ignoredLiteral -> {
-                // Constants do not contribute referenced variables.
-            }
-
-            case ConstraintAst ignored -> {
-                // Other constraint shapes are ignored until they are supported here.
-            }
-        }
+    private String buildOutOfScopeVariableMessage(String variableName, String clause) {
+        return "Variable ?" + variableName + " used in " + clause + " is not visible in WHERE clause";
     }
 
     /**
